requirements.txt
torch==2.6.0 torchvision==0.21.0 torchaudio==2.6.0 --index-url https://download.pytorch.org/whl/cu124
numpy
scikit-learn
h5py
tqdm
scipy
matplotlib
joblib
pyyaml
torchdata
git+https://github.com/nicola-decao/power_spherical.git#egg=power_spherical
git+https://github.com/pafoster/pyitlib.git@master